+++ Corelayer/DTOs/Comment/CreateCommentDto.cs
@@ -8,16 +8,16 @@ namespace CoreLayer.DTOs.Comment
 {
     public class CreateCommentDto
     {
-        public long UserId { get; set; }
-        public long ProductId { get; set; }
+        public int UserId { get; set; }
+        public int ProductId { get; set; }
         public string Text { get; set; }
     }
 
     public class CommentDto
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public string UserFullName { get; set; }
-        public long ProductId { get; set; }
+        public int ProductId { get; set; }
         public string Text { get; set; }
         public DateTime DateTime { get; set; }
