+++ Corelayer/DTOs/Product/ProductDto.cs
@@ -5,12 +5,12 @@ namespace CoreLayer.DTOs.Product;
 
 public class ProductDto
 {
-    public long Id { get; set; }
+    public int Id { get; set; }
     public string Title { get; set; }
     public string ImageName { get; set; }
     public string Description { get; set; }
-    public long CategoryId { get; set; }
-    public long? SubCategoryId { get; set; }
+    public int CategoryId { get; set; }
+    public int? SubCategoryId { get; set; }
     public string Slug { get; set; }
     public string SeoData { get; set; }
     public int Count { get; set; }
@@ -28,8 +28,8 @@ public class CreateProductDto
     public string Title { get; set; }
     public IFormFile ImageFile { get; set; }
     public string Description { get; set; }
-    public long CategoryId { get; set; }
-    public long? SubCategoryId { get; set; }
+    public int CategoryId { get; set; }
+    public int? SubCategoryId { get; set; }
     public string Slug { get; set; }
     public string SeoData { get; set; }
     public int Count { get; set; }
@@ -41,12 +41,12 @@ public class CreateProductDto
 }
 public class EditProductDto
 {
-    public long ProductId { get; set; }
+    public int ProductId { get; set; }
     public string Title { get; set; }
     public IFormFile ImageFile { get; set; }
     public string Description { get; set; }
-    public long CategoryId { get; set; }
-    public long? SubCategoryId { get; set; }
+    public int CategoryId { get; set; }
+    public int? SubCategoryId { get; set; }
     public string Slug { get; set; }
     public string SeoData { get; set; }
     public int Count { get; set; }
