+++ Corelayer/DTOs/Product/ProductImage/AddImageDto.cs
@@ -5,14 +5,14 @@ namespace CoreLayer.DTOs.Product.ProductImage;
 
 public class AddImageDto
 {
-    public long ProductId { get; set; }
+    public int ProductId { get; set; }
     public IFormFile ImageFile { get; set; }
 }
 
 public class ImageDto
 {
-    public long Id { get; set; }
-    public long ProductId { get; set; }
+    public int Id { get; set; }
+    public int ProductId { get; set; }
     public string ImageName { get; set; }
     public DataLayer.Entities.Products.Product Product { get; set; }
 }
