+++ Corelayer/Utilities/UserUtil.cs
@@ -9,12 +9,12 @@ namespace CoreLayer.Utilities
 {
     public static class UserUtil
     {
-        public static long GetUserId(this ClaimsPrincipal principal)
+        public static int GetUserId(this ClaimsPrincipal principal)
         {
             if (principal == null)
                 throw new ArgumentNullException(nameof(principal));
 
-            return Convert.ToInt64(principal.FindFirst(ClaimTypes.NameIdentifier)?.Value);
+            return Convert.ToInt32(principal.FindFirst(ClaimTypes.NameIdentifier)?.Value);
         }
         public static string? GetNameUser(this ClaimsPrincipal principal)
         {
